Modularize riscv-macros code
